requirements.txt
requests
playwright
python-docx
